+++ bin/select-references.py
@@ -2,7 +2,7 @@
 """
 """
 PROGRAM = "select-references"
-VERSION = "1.1.0"
+VERSION = "1.2.0"
 
 if __name__ == '__main__':
     import argparse as ap